feat(architecture): architecture first review and refactor
