+++ src/components/commingSoon/commingSoon.jsx
@@ -1,5 +1,5 @@
 import React, { Fragment } from 'react';
-import Scroll from './Scroll';
+import Scroll from '../utils/Scroll';
 
 const CoomingSoon = () => {
   return (
@@ -7,6 +7,7 @@ const CoomingSoon = () => {
       <section id="banner">
         <div className="inner">
           <h2>COMING SOON</h2>
+          <p>María José Méndez</p>
           <p>FullStack Development | Software Development | Web Design</p>
           <p>mj@mtekh.com</p>
           <ul className="actions special"></ul>
