+++ src/components/utils/Scroll.jsx
@@ -1,4 +1,4 @@
-import smoothscroll from 'smoothscroll-polyfill';
+import smoothScroll from 'smoothscroll-polyfill';
 import React from 'react';
 import PropTypes from 'prop-types';
 
@@ -19,7 +19,7 @@ class Scroll extends React.Component {
     this.handleClick = this.handleClick.bind(this);
   }
   componentDidMount() {
-    smoothscroll.polyfill();
+    smoothScroll.polyfill();
   }
   handleClick(e) {
     e.preventDefault();
