+++ src/pages/Generic.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Layout from '../components/Layout';
+import Layout from '../components/layout/Layout';
 
 const IndexPage = () => (
   <Layout fullMenu>
